Move datatypes to core
