+++ core/specfem/algorithms/divergence.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "datatype/point_view.hpp"
 #include "execution/for_each_level.hpp"
+#include "specfem/datatype.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_Core.hpp>
