+++ core/specfem/chunk_element/stress_integrand.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "datatype/chunk_element_view.hpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
+#include "specfem/datatype.hpp"
 #include <Kokkos_Core.hpp>
 
 namespace specfem {
