+++ core/specfem/data_access/accessor/chunk_edge.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "datatype/simd.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem/data_access/accessor.hpp"
+#include "specfem/datatype.hpp"
 #include <Kokkos_Core.hpp>
 
 namespace specfem::data_access {
