+++ core/specfem/data_access/accessor/point_accessor.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "datatype/interface.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem/data_access/accessor.hpp"
+#include "specfem/datatype.hpp"
 
 namespace specfem::data_access {
