+++ core/specfem/datatype/simd.hpp
