+++ core/specfem/point/field_derivatives.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "datatype/interface.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem/data_access.hpp"
+#include "specfem/datatype.hpp"
 #include <Kokkos_Core.hpp>
 
 namespace specfem {
