+++ docs/sections/api/index.rst
@@ -29,7 +29,6 @@ Old structure
     enumerations/index
     quadrature/index
     mesh/index
-    datatype/index
     execution/index
     medium/index
     kokkos_kernels/index
