+++ docs/sections/api/specfem/datatype/impl_chunk_edge_subview.rst
