+++ docs/sections/api/specfem/datatype/impl_register_array.rst
