+++ docs/sections/api/specfem/datatype/scalar_chunk_edge_view.rst
