+++ docs/sections/api/specfem/datatype/simd.rst
