+++ docs/sections/api/specfem/datatype/tensor_chunk_edge_view.rst
