+++ docs/sections/api/specfem/datatype/vector_chunk_edge_view.rst
