+++ include/kokkos_kernels/impl/compute_mass_matrix.tpp
@@ -2,7 +2,7 @@
 
 #include "specfem/boundary_conditions.hpp"
 #include "specfem/boundary_conditions.tpp"
-#include "datatype/interface.hpp"
+#include "specfem/datatype.hpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
 #include "enumerations/wavefield.hpp"
