+++ include/kokkos_kernels/impl/compute_stiffness_interaction.tpp
@@ -3,7 +3,7 @@
 #include "specfem/algorithms.hpp"
 #include "specfem/boundary_conditions.hpp"
 #include "specfem/boundary_conditions.tpp"
-#include "datatype/interface.hpp"
+#include "specfem/datatype.hpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
 #include "enumerations/wavefield.hpp"
