+++ include/quadrature/lagrange_derivative.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "datatype/interface.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem/data_access.hpp"
+#include "specfem/datatype.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
