+++ tests/unit-tests/algorithms/dim2/gradient.cpp
@@ -28,8 +28,6 @@
 #include <vector>
 
 #include "SPECFEM_Environment.hpp"
-#include "datatype/point_view.hpp"
-#include "datatype/simd.hpp"
 #include "enumerations/interface.hpp"
 #include "execution/chunked_domain_iterator.hpp"
 #include "execution/for_each_level.hpp"
@@ -37,6 +35,7 @@
 #include "quadrature/interface.hpp"
 #include "specfem/algorithms/gradient.hpp"
 #include "specfem/assembly.hpp"
+#include "specfem/datatype.hpp"
 
 namespace specfem::algorithms_test {
