+++ tests/unit-tests/algorithms/dim2/transfer.cpp
@@ -7,11 +7,11 @@
 #include <Kokkos_Core.hpp>
 #include <gtest/gtest.h>
 
-#include "datatype/chunk_edge_view.hpp"
 #include "enumerations/interface.hpp"
 #include "specfem/algorithms/transfer.hpp"
 #include "specfem/chunk_edge.hpp"
 #include "specfem/data_access.hpp"
+#include "specfem/datatype.hpp"
 #include "utilities/include/fixture/nonconforming_interface.hpp"
 #include "utilities/interface.hpp"
