+++ tests/unit-tests/assembly/kernels/kernels.cpp
@@ -1,8 +1,8 @@
 #include "../test_fixture/test_fixture.hpp"
-#include "datatype/simd.hpp"
 #include "enumerations/dimension.hpp"
 #include "execution/chunked_domain_iterator.hpp"
 #include "execution/for_all.hpp"
+#include "specfem/datatype.hpp"
 #include "specfem/macros.hpp"
 #include "specfem_setup.hpp"
 #include <gtest/gtest.h>
