+++ tests/unit-tests/assembly/properties/properties.cpp
@@ -1,11 +1,11 @@
 #include "../test_fixture/test_fixture.hpp"
-#include "datatype/simd.hpp"
 #include "enumerations/dimension.hpp"
 #include "execution/chunked_domain_iterator.hpp"
 #include "execution/for_all.hpp"
 #include "io/ASCII/ASCII.hpp"
 #include "io/property/reader.hpp"
 #include "io/property/writer.hpp"
+#include "specfem/datatype.hpp"
 #include "specfem/macros.hpp"
 #include "specfem_setup.hpp"
 #include <gtest/gtest.h>
