+++ tests/unit-tests/point/kernels/kernels_tests.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "datatype/simd.hpp"
+#include "specfem/datatype.hpp"
 #include "specfem_setup.hpp"
 #include "test_helper.hpp"
 #include <Kokkos_Core.hpp>
